fixup! chore(suite): small SuiteLayout refactor
