+++ packages/suite/src/hooks/suite/useClearAnchorHighlightOnClick.ts
